Moved weapon scripts
