+++ BNA_KC_Scripts/Data/Functions/EMP/fn_disableDekaShields.sqf
