+++ BNA_KC_Scripts/Data/Functions/EMP/fn_tempDisableVehicles.sqf
