+++ BNA_KC_Scripts/Data/Functions/Settings/fn_weapOptions.sqf
