+++ BNA_KC_Scripts/Data/Functions/Weapons/fn_specialAmmo.sqf
