+++ BNA_KC_Scripts/config.cpp
@@ -56,6 +56,21 @@ class CfgFunctions
             class CreateMarkAction{};
         };
 
+        class EMP
+        {
+            file = "BNA_KC_Scripts\Data\Functions\EMP";
+            class killDroids;
+            class disableDekaShields;
+            class tempDisableVehicles;
+            class playDroidPopperSound;
+        };
+
+        class Weapons
+        {
+            file = "BNA_KC_Scripts\Data\Functions\Weapons";
+            class specialAmmo;
+        };
+
         class Medical
         {
             file = "BNA_KC_Scripts\Data\Functions\Medical";
@@ -75,6 +90,10 @@ class CfgFunctions
             {
                 file = "BNA_KC_Scripts\Data\Functions\Settings\fn_miscKeybinds.sqf";
             };
+            class weapOptions
+            {
+                file = "BNA_KC_Scripts\Data\Functions\Settings\fn_weapOptions.sqf";
+            };
         };
 	};
 };
@@ -84,6 +103,10 @@ class Extended_PreInit_EventHandlers
     class BNA_KC_Misc_SettingsPreInit
     {
         init = "call BNAKC_fnc_miscOptions; call BNAKC_fnc_miscKeybinds;";
+    };
+    class BNA_KC_Weap_OptionsPreInit
+    {
+        init = "call BNAKC_fnc_weapOptions;";
     };
 };
 
@@ -104,6 +127,10 @@ class Extended_PostInit_EventHandlers
     class BNA_KC_Scripts_FixInfiniteLoading
     {
         init = "if (BNA_KC_Misc_FixInfiniteLoading) then { call BNAKC_fnc_FixInfiniteLoading; };";
+    };
+    class BNA_KC_Weap_SpecialAmmo
+    {
+        init = "['ace_firedPlayer'] call BNAKC_fnc_specialAmmo;";
     };
 };
